+++ lib/rpcs_dpdk/ethdev.c
@@ -49,19 +49,19 @@
 #define CASE_RTE2TARPC(_rte) \
     case (_rte): *rpc = TARPC_##_rte; break
 
-static uint32_t
-tarpc_rte_rx_offloads2rpc(uint32_t rte)
+static uint64_t
+tarpc_rte_rx_offloads2rpc(uint64_t rte)
 {
-    uint32_t    rpc = 0;
+    uint64_t rpc = 0;
 
 #define RTE_DEV_RX_OFFLOAD2RPC(_bit) \
     do {                                                            \
-        uint32_t flag = DEV_RX_OFFLOAD_##_bit;                      \
+        uint64_t flag = DEV_RX_OFFLOAD_##_bit;                      \
                                                                     \
         if (rte & flag)                                             \
         {                                                           \
             rte &= ~flag;                                           \
-            rpc |= (1 << TARPC_RTE_DEV_RX_OFFLOAD_##_bit##_BIT);    \
+            rpc |= (1ULL << TARPC_RTE_DEV_RX_OFFLOAD_##_bit##_BIT); \
         }                                                           \
     } while (0)
     RTE_DEV_RX_OFFLOAD2RPC(VLAN_STRIP);
@@ -73,7 +73,7 @@ tarpc_rte_rx_offloads2rpc(uint32_t rte)
     RTE_DEV_RX_OFFLOAD2RPC(OUTER_IPV4_CKSUM);
 #undef RTE_DEV_RX_OFFLOAD2RPC
     if (rte != 0)
-        rpc |= (1 << TARPC_RTE_DEV_RX_OFFLOAD__UNKNOWN_BIT);
+        rpc |= (1ULL << TARPC_RTE_DEV_RX_OFFLOAD__UNKNOWN_BIT);
     return rpc;
 }
